Update patch set 1
Patch Set 1:

We have the problem on endeavoru that scanning for new APs doesn't work when you are associated with one. You can check that with an app like WiFi Analyzer. You don't see any AP or you only see the WiFi you're connected to and the signal strength doesn't change.

The problem arises due to a change in the framework that was introduced with Android 4.3. The framework now sends a "SCAN
TYPE=ONLY" command (previously just "SCAN") to the wpa_supplicant which doesn't recognize it. This patch changes that.

If my understanding is correct then this problem should affect all devices using this supplicant since CM 10.2. So you should test if scanning works when connected to an AP, without and with the patch.

Unfortunately there is now a problem where it can happen that the scanned list of AP gets stuck after some time. I don't know what the problem is and if it is related to the wpa_supplicant. It may have been there in CM 10.1 and may be specific to us.
It only happens to me when I'm at work where there are a lot more APs. You can see this nicely in the WiFi analyzer app because all signal strength lines are now perfectly flat and don't fluctuate anymore.
Unfortunately I cannot properly debug this at work :(.

So might want to watch out for the above also.

Anyway I think this patch is a step in the right direction if it doesn't break thinks. And if somebody knows other devices that might affected please add the developers.

Patch-set: 1