Update patch set 1
Patch Set 1: Code-Review+1

Tested this patch on I9003 (OMAP3630) with TI WL1271 WiFI + BT Combo Chip.

Scanning works fine when associated with the AP. Earlier, it wasn't working. However, as mentioned in the comments, it gets stuck sometimes. We are using the latest R5.SP4 wl12xx driver from https://github.com/TI-OpenLink

It seems that not many devices use this supplicant. This supplicant is only required for hostapd support on TI chipset like that on I9003 and endeavoru. If there is no need of hostapd (Required for Tethering), the upstream supplicant could do the job well. A similar device to I9003, encore uses the generic supplicant even when it has WL1271 chip similar to I9003, but doesn't required hostapd because of it being a WiFi only tablet.
https://github.com/CyanogenMod/android_device_bn_encore/blob/cm-10.2/BoardConfig.mk#L103

Patch-set: 1
Label: Code-Review=+1